backend/requirements.txt
# Porakhela Backend Requirements

# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# File Upload & Media
Pillow==10.1.0
django-storages==1.14.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==19.12.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# HTTP Client for Applink APIs
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.0
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0